+++ src/blueetl/extract/simulations.py
@@ -205,16 +205,23 @@ class Simulations(BaseExtractor):
     def from_config(cls, config: SimulationsConfig, query: Optional[dict] = None) -> "Simulations":
         """Extract simulations from the given simulation campaign."""
         df = config.to_pandas()
+        original_len = len(df)
         df = cls._from_paths(df)
         df = cls._filter_simulations_df(df, query)
-        return cls(df)
+        return cls(df, cached=False, filtered=len(df) != original_len)
 
     @classmethod
-    def from_pandas(cls, df: pd.DataFrame, query: Optional[dict] = None) -> "Simulations":
+    def from_pandas(
+        cls,
+        df: pd.DataFrame,
+        query: Optional[dict] = None,
+        cached: bool = True,
+    ) -> "Simulations":
         """Extract simulations from a dataframe containing valid simulation ids and circuit ids."""
+        original_len = len(df)
         df = cls._from_paths(df)
         df = cls._filter_simulations_df(df, query, cached=True)
-        return cls(df)
+        return cls(df, cached=cached, filtered=len(df) != original_len)
 
     def to_pandas(self) -> pd.DataFrame:
         """Dump simulations to a dataframe that can be serialized and stored."""