+++ tests/unit/extract/test_simulations.py
@@ -495,7 +495,7 @@ def test_simulations_to_pandas():
             },
         ]
     )
-    result_df = test_module.Simulations(df).to_pandas()
+    result_df = test_module.Simulations(df, cached=False, filtered=False).to_pandas()
     expected_df = pd.DataFrame(
         [
             {