+++ CHANGELOG.rst
@@ -11,7 +11,10 @@ New Features
 - Support combination of parameters in features configuration [NSETM-2091]
 - Add analysis configuration model and validation [NSETM-2099]
 - Add blueetl CLI [NSETM-2115]
+- Improve performance of report extraction [NSETM-2116]
 - Improve performance of features calculation [NSETM-2116]
+- Process features in group when possible.
+- Add `cached` and `filtered` attributes to `BaseExtractor`.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~